add migration to update existing actors
